Collocate all apply code
